[od4.6] Another fix in compiler warning
Summary: Fixing a silly mistake in D3033

Test Plan: Only CDash or windows giys have the authority to test

Reviewers: ranojay

Reviewed By: ranojay

CC: raman

Differential Revision: https://phabricator.dgbes.com:8443/D3035

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@33199 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/uiTools/uisaveimagedlg.cc
@@ -512,7 +512,7 @@ void uiSaveImageDlg::setSizeInPix( int width, int height )
 {
     sizepix_.setWidth( mCast(float,width) );
     sizepix_.setHeight( mCast(float,height) );
-    const float dpi = mCast(int,dpifld_->box()->getValue());
+    const float dpi = mCast(float,dpifld_->box()->getValue());
     sPixels2Inch( sizepix_, sizeinch_, dpi );
     sInch2Cm( sizeinch_, sizecm_ );
     unitChg( 0 );
